% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{add_dimension_instances}
\alias{add_dimension_instances}
\title{Add dimension instances}
\usage{
add_dimension_instances(db, name, table)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}

\item{table}{A table of new instances.}
}
\value{
A \code{star_database} object.
}
\description{
Add dimension instances
}
\keyword{internal}
