% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{add_dput_column}
\alias{add_dput_column}
\title{For each row, add a vector of values}
\usage{
add_dput_column(v, column)
}
\arguments{
\item{v}{A \code{tibble}, rows of a dimension table.}

\item{column}{A string, name of the column to include a vector of values.}
}
\value{
A \code{tibble}, rows of a dimension table.
}
\description{
For each row, add a vector of values
}
\keyword{internal}
