% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_operation}
\alias{add_operation}
\title{A data frame row is added with a new operation}
\usage{
add_operation(op = NULL, op_name, name = "", details = NULL, details2 = NULL)
}
\arguments{
\item{op}{A data frame of operations.}

\item{op_name}{A string, operation name.}

\item{name}{A string, element name.}

\item{details}{A vector of strings, operation details.}

\item{details2}{A vector of strings, operation additional details.}
}
\value{
A data frame.
}
\description{
A data frame row is added with a new operation
}
\keyword{internal}
