% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{add_surrogate_key.dimension_table}
\alias{add_surrogate_key.dimension_table}
\title{Add the surrogate key from a dimension table to the instances table.}
\usage{
\method{add_surrogate_key}{dimension_table}(dimension_table, instances)
}
\arguments{
\item{dimension_table}{A \code{dimension_table} object.}

\item{instances}{A \code{tibble}, the instances table.}
}
\value{
A \code{tibble}.
}
\description{
Add the surrogate key from a dimension table to the instances table.
}
\keyword{internal}
