% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_table.R
\name{add_surrogate_key}
\alias{add_surrogate_key}
\title{Add the surrogate key from a dimension table to the instances table.}
\usage{
add_surrogate_key(instances, dimension, surrogate_key)
}
\arguments{
\item{instances}{A \code{tibble}, the instances table.}

\item{dimension}{A \code{tibble}, the dimension table.}

\item{surrogate_key}{A string, the surrogate key name.}
}
\value{
A \code{tibble}.
}
\description{
Add the surrogate key from a dimension table to the instances table.
}
\keyword{internal}
