% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{apply_filter_dimension}
\alias{apply_filter_dimension}
\title{Apply filter dimension}
\usage{
apply_filter_dimension(db, sq)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{sq}{A \code{star_query} object.}
}
\description{
Select the instances of the dimensions that meet the defined conditions.
}
\keyword{internal}
