% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{apply_select_dimension}
\alias{apply_select_dimension}
\title{Apply select dimension}
\usage{
apply_select_dimension(db, sq)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{sq}{A \code{star_query} object.}
}
\description{
Select dimensions and attributes.
}
\keyword{internal}
