% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{apply_select_fact}
\alias{apply_select_fact}
\title{Apply select fact}
\usage{
apply_select_fact(db, sq)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{sq}{A \code{star_query} object.}
}
\description{
Select the facts, measures and define the aggregation functions.
}
\keyword{internal}
