% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{as_dm_class}
\alias{as_dm_class}
\title{Generate a \code{dm} class with fact and dimension tables}
\usage{
as_dm_class(db)
}
\arguments{
\item{db}{A \code{star_database} or \code{constellation} object.}
}
\value{
A \code{dm} object.
}
\description{
To port databases to other work environments it is useful to be able to
export them as a \code{dm} class, as this function does, in this way it can be
saved directly in a DBMS.
}
\examples{

db1 <- star_database(mrs_cause_schema, ft_num) |>
  snake_case()
dm1 <- db1 |>
  as_dm_class()

db2 <- star_database(mrs_age_schema, ft_age) |>
  snake_case()

ct <- constellation("MRS", list(db1, db2))
dm <- ct |>
  as_dm_class()

}
\seealso{
\code{\link{mrs_cause_schema}}, \code{\link{mrs_age_schema}}, \code{\link{ft_num}}, \code{\link{ft_age}}

Other star database definition functions: 
\code{\link{as_tibble_list}()},
\code{\link{constellation}()},
\code{\link{snake_case}()},
\code{\link{star_database}()}
}
\concept{star database definition functions}
