% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{as_tibble_list.constellation}
\alias{as_tibble_list.constellation}
\title{Generate a list of tibbles with fact and dimension tables}
\usage{
\method{as_tibble_list}{constellation}(db)
}
\arguments{
\item{db}{A \code{constellation} object.}
}
\value{
A list of \code{tibble} objects.
}
\description{
To port databases to other work environments it is useful to be able to
export them as a list of tibbles, as this function does.
}
\examples{

db1 <- star_database(mrs_cause_schema, ft_num) |>
  snake_case()
db2 <- star_database(mrs_age_schema, ft_age) |>
  snake_case()
ct <- constellation("MRS", list(db1, db2))
tl <- ct |>
  as_tibble_list()

}
