% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R, R/generic.R, R/star_database.R
\name{as_tibble_list.constellation}
\alias{as_tibble_list.constellation}
\alias{as_tibble_list}
\alias{as_tibble_list.star_database}
\title{Generate a list of tibbles with fact and dimension tables}
\usage{
\method{as_tibble_list}{constellation}(db)

as_tibble_list(db)

\method{as_tibble_list}{star_database}(db)
}
\arguments{
\item{db}{A \code{star_database} or \code{constellation} object.}
}
\value{
A list of \code{tibble}
}
\description{
To port databases to other work environments it is useful to be able to
export them as a list of tibbles, as this function does.
}
\examples{

s <- star_schema() |>
  define_facts(fact_schema(
    name = "mrs_cause",
    measures = c(
      "Pneumonia and Influenza Deaths",
      "All Deaths"
    )
  )) |>
  define_dimension(dimension_schema(
    name = "when",
    attributes = c(
      "Year"
    )
  )) |>
  define_dimension(dimension_schema(
    name = "where",
    attributes = c(
      "REGION",
      "State",
      "City"
    )
  ))

# ft_num contains instances
db <- star_database(s, ft_num) |>
  snake_case()

tl <- db |>
  as_tibble_list()

}
\seealso{
Other star schema and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{snake_case}()}
}
\concept{star schema and constellation definition functions}
\keyword{internal}
