% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{as_tibble_list.star_database}
\alias{as_tibble_list.star_database}
\title{Generate a list of tibbles with fact and dimension tables}
\usage{
\method{as_tibble_list}{star_database}(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A list of \code{tibble}
}
\description{
To port databases to other work environments it is useful to be able to
export them as a list of tibbles, as this function does.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num) |>
  snake_case()

tl <- db |>
  as_tibble_list()

}
