% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{as_tibble_list_common}
\alias{as_tibble_list_common}
\title{Generate a list of tibbles with fact and dimension tables}
\usage{
as_tibble_list_common(dimensions, facts)
}
\arguments{
\item{dimensions}{A list of dimension tables.}

\item{facts}{A list of fact tables.}
}
\value{
A list of \code{tibble} objects.
}
\description{
Generate a list of tibbles with fact and dimension tables
}
\keyword{internal}
