% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{check_refesh}
\alias{check_refesh}
\alias{check_refesh.star_database}
\title{Checks the refresh of the selected star database from the given database}
\usage{
check_refesh(db, star, refresh_db)

\method{check_refesh}{star_database}(db, star = 1, refresh_db)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{star}{A string, star database name or index in constellation, database
to refresh.}

\item{refresh_db}{A \code{star_database} object with the same structure with
new data.}
}
\value{
A \code{star_database} object.
}
\description{
Checks the refresh operation of the selected star database from the given
database. Once this operation is carried out, the results can be consulted on
the new instances in dimensions or existing instances in the facts.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num) |>
  check_refesh()

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}
}
\concept{star database transformation functions}
