% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{check_schema_validity}
\alias{check_schema_validity}
\title{check if star schema is valid: has facts and dimensions}
\usage{
check_schema_validity(schema)
}
\arguments{
\item{schema}{A \code{star_schema} object.}
}
\value{
A boolean.
}
\description{
check if star schema is valid: has facts and dimensions
}
\keyword{internal}
