% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{conform_dimensions}
\alias{conform_dimensions}
\title{Conform dimensions}
\usage{
conform_dimensions(to_conform)
}
\arguments{
\item{to_conform}{A \code{dimension_table} object list.}
}
\value{
A \code{dimension_table} object.
}
\description{
Generate a dimension from a list of dimensions with the same schema.
}
\keyword{internal}
