% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{constellation}
\alias{constellation}
\title{\code{constellation} S3 class}
\usage{
constellation(name = NULL, stars = NULL)
}
\arguments{
\item{name}{A string.}

\item{stars}{A list of \code{star_database} objects.}
}
\value{
A \code{constellation} object.
}
\description{
Creates a \code{constellation} object from a list of \code{star_database} objects. All
dimensions with the same name in the star schemas have to be conformable
(share the same structure, even though they have different instances).
}
\seealso{
Other star schema and constellation definition functions: 
\code{\link{snake_case}()}
}
\concept{star schema and constellation definition functions}
