% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_finest}
\alias{db_finest}
\title{Czech debit card company specialising on payments at gas stations (finest detail)}
\format{
A \code{star_database}.
}
\source{
\url{https://relational.fit.cvut.cz/dataset/CCS}
}
\usage{
db_finest
}
\description{
Multidimensional design with finest detail from the data available at the source.
}
\details{
The development has been done in the vignette \code{vignette("v20-rdbms-dm")}.
}
\seealso{
\code{\link{mrs_cause_schema}}

Other database example data: 
\code{\link{db_summary}}
}
\concept{database example data}
\keyword{datasets}
