% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_finest_dm}
\alias{db_finest_dm}
\title{Czech debit card company specialising on payments at gas stations (finest granularity)}
\format{
A \code{star_database}.
}
\source{
\url{https://relational.fit.cvut.cz/dataset/CCS}
}
\usage{
db_finest_dm
}
\description{
Multidimensional design with the finest granularity possible from the data
available at the source.
}
\details{
The development has been done in the vignette \code{vignette("dm")} of the \code{rolap}
package.
}
\seealso{
\code{\link{mrs_cause_schema}}

Other database example data: 
\code{\link{db_summary_dm}}
}
\concept{database example data}
\keyword{datasets}
