% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{default_disconnect}
\alias{default_disconnect}
\title{Default disconnect function}
\usage{
default_disconnect(con)
}
\arguments{
\item{con}{A \code{DBI::DBIConnection} object.}
}
\value{
TRUE, invisibly.
}
\description{
Generate sql code for the first refresh operation.
}
\keyword{internal}
