% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{define_facts}
\alias{define_facts}
\title{Define facts in a \code{star_schema} object.}
\usage{
define_facts(star, facts)
}
\arguments{
\item{star}{A \code{star_schema} object.}

\item{facts}{A \code{fact_schema} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Define facts in a \code{star_schema} object.
}
\examples{

s <- star_schema()
f <- fact_schema(
  name = "mrs_cause",
  measures = c(
    "Pneumonia and Influenza Deaths",
    "Other Deaths"
  )
)
s <- s |>
  define_facts(f)
}
\seealso{
\code{\link{star_schema}}

Other star schema definition functions: 
\code{\link{fact_schema}()},
\code{\link{star_schema}()}
}
\concept{star schema definition functions}
