% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{delete_all_operations_found}
\alias{delete_all_operations_found}
\title{Delete in stars all operations found}
\usage{
delete_all_operations_found(stars, op)
}
\arguments{
\item{stars}{A list of \code{star_database} objects.}

\item{op}{A \code{star_operations} object.}
}
\value{
A list of \code{star_database} objects.
}
\description{
Delete in stars all operations found
}
\keyword{internal}
