% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_operation}
\alias{delete_operation}
\title{Delete an operation}
\usage{
delete_operation(op, op_name, name = NULL, details = NULL, details2 = NULL)
}
\arguments{
\item{op}{A \code{star_operation} object.}

\item{op_name}{A string, operation name.}

\item{name}{A string, element name.}

\item{details}{A vector of strings, operation details.}

\item{details2}{A vector of strings, operation additional details.}
}
\value{
op A \code{star_operation} object.
}
\description{
Delete an operation
}
\keyword{internal}
