% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database_export.R
\name{draw_tables}
\alias{draw_tables}
\alias{draw_tables.star_database}
\title{Draw tables}
\usage{
draw_tables(db)

\method{draw_tables}{star_database}(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
An object with a \code{print()} method.
}
\description{
Draw the tables of the ROLAP star diagrams.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num) |>
  snake_case()
\donttest{
db |>
  draw_tables()
}

}
\seealso{
\code{\link{star_database}}

Other star database exportation functions: 
\code{\link{as_csv_files}()},
\code{\link{as_dm_class}()},
\code{\link{as_rdb}()},
\code{\link{as_single_tibble_list}()},
\code{\link{as_tibble_list}()},
\code{\link{as_xlsx_file}()}
}
\concept{star database exportation functions}
