% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolap.R
\name{explore_schema}
\alias{explore_schema}
\title{Explore MS OLAP Schema}
\usage{
explore_schema(con, field, clean_names = FALSE)
}
\arguments{
\item{con}{A list of connection parameters, as returned by 'Rolap::olap_connection'.}

\item{field}{The field of information to explore. Read more in details.}

\item{clean_names}{Clean column names by the use of 'janitor::clean_names'?}
}
\description{
This function returns a tibble of information regarding the available Cubes
or the contents of Cubes.
}
\details{
\subsection{Fields}{

The schema of information from the database contains different fields of information.
You need to specify which type of information you want to explore. Usually,
you should get most of the useful information from the following filelds: 'Cubes', 'Hierarchies',
'Levels' and 'Measures'. All available fields to explore can be found
\href{https://docs.microsoft.com/en-us/dotnet/api/microsoft.analysisservices.adomdclient.adomdschemaguid?view=analysisservices-dotnet#fields}{here}.
}
}
\examples{
\dontrun{
con <- olap_connection("user", "pwd", "https://myolapcube.mycompany.com/msolap/", catalog = "Sales")
explore_schema(con, "Levels")
 }
}
\keyword{MDX}
\keyword{olap}
