% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_schema.R
\name{fact_schema}
\alias{fact_schema}
\title{\code{fact_schema} S3 class}
\usage{
fact_schema(name = NULL, measures = NULL)
}
\arguments{
\item{name}{A string, name of the fact.}

\item{measures}{A vector of measure names.}
}
\value{
A \code{fact_schema} object.
}
\description{
A \code{fact_schema} object is created, the essential data is a name and
a set of measurements that can be empty (does not have explicit
measurements).
}
\details{
A \code{fact_schema} object is part of a \code{star_schema} object, defines
the facts of the star schema.
}
\examples{

f <- fact_schema(
  name = "mrs_cause",
  measures = c(
    "Pneumonia and Influenza Deaths",
    "Other Deaths"
  )
)

}
\seealso{
\code{\link{star_schema}}

Other star schema definition functions: 
\code{\link{define_dimension}()},
\code{\link{define_facts}()},
\code{\link{dimension_schema}()},
\code{\link{star_schema}()}
}
\concept{star schema definition functions}
