% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table.R
\name{fact_table}
\alias{fact_table}
\title{\code{fact_table} S3 class}
\usage{
fact_table(fact_schema = NULL, instances = NULL, surrogate_keys)
}
\arguments{
\item{fact_schema}{A \code{dimension_schema} object.}

\item{instances}{A flat table with the fact instances.}

\item{surrogate_keys}{A vector of strings, surrogate key names.}
}
\value{
A \code{fact_table} object.
}
\description{
A \code{fact_table} object is created, we have to get its
surrogate keys.
}
\keyword{internal}
