% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{filter_geo_attributes}
\alias{filter_geo_attributes}
\title{From attributes, leave only these contained in dimensions}
\usage{
filter_geo_attributes(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A list of geodimensions.
}
\description{
From attributes, leave only these contained in dimensions
}
\keyword{internal}
