% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{filter_geo_dimensions}
\alias{filter_geo_dimensions}
\title{From geodimensions, leave only contained in vector of names}
\usage{
filter_geo_dimensions(db, dim)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{names}{A vector of strings, dimension names.}
}
\value{
A list of geodimensions.
}
\description{
From geodimensions, leave only contained in vector of names
}
\keyword{internal}
