% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_playing.R
\name{filter_rpd_dimensions}
\alias{filter_rpd_dimensions}
\title{From rpd dimensions, leave only contained in vector of names.}
\usage{
filter_rpd_dimensions(db, names)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{names}{A vector of strings, dimension names.}
}
\value{
A list of vectors of dimension names.
}
\description{
From rpd dimensions, leave only contained in vector of names.
}
\keyword{internal}
