% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{flat_table}
\alias{flat_table}
\title{\code{flat_table} S3 class}
\usage{
flat_table(name = NULL, instances, unknown_value = NULL)
}
\arguments{
\item{name}{A string.}

\item{instances}{A \code{tibble}, table of instances.}

\item{unknown_value}{A string, value used to replace empty and NA values in
attributes.}
}
\value{
A \code{flat_table} object.
}
\description{
Creates a \code{flat_table} object.
}
\details{
The objective is to allow the transformation of flat tables.

We can also indicate the value that is used in the data with undefined values.
}
\examples{

ft <- flat_table('iris', iris)

ft <- flat_table('ft_num', ft_num)

}
\seealso{
\code{\link{star_database}}

Other flat table definition functions: 
\code{\link{read_flat_table_file}()},
\code{\link{read_flat_table_folder}()},
\code{\link{select_attributes}()},
\code{\link{select_measures}()}
}
\concept{flat table definition functions}
