% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_cause_rpd}
\alias{ft_cause_rpd}
\title{Mortality Reporting System by Cause}
\format{
A \code{tibble}.
}
\source{
\url{https://catalog.data.gov/dataset/deaths-in-122-u-s-cities-1962-2016-122-cities-mortality-reporting-system}
}
\usage{
ft_cause_rpd
}
\description{
Selection of data from the 122 Cities Mortality Reporting System by cause,
for the first 9 weeks of 1962 and 4 cities.
}
\details{
The original dataset begins in 1962. For each week, in 122 US cities,
mortality figures by age group and cause, considered separately, are included
(i.e., the combination of age group and cause is not included). In the cause,
only a distinction is made between pneumonia or influenza and others.

Two additional dates have been generated, which were not present in the
original dataset.
}
\seealso{
\code{\link{mrs_cause_schema}}

Other mrs example data: 
\code{\link{ft_age_rpd}},
\code{\link{ft_age}},
\code{\link{ft_num}},
\code{\link{ft}},
\code{\link{mrs_ft}}
}
\concept{mrs example data}
\keyword{datasets}
