% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{generate_table_sql}
\alias{generate_table_sql}
\title{Generate table sql}
\usage{
generate_table_sql(table)
}
\arguments{
\item{table}{A \code{tibble}.}
}
\value{
A string.
}
\description{
Generate sql code for a table.
}
\keyword{internal}
