% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{generate_table_sql_delete}
\alias{generate_table_sql_delete}
\title{Generate table sql delete}
\usage{
generate_table_sql_delete(table, instances)
}
\arguments{
\item{table}{A string, table name.}

\item{instances}{A \code{tibble}.}
}
\value{
A vector of strings.
}
\description{
Generate sql code for deleting instances in a table.
}
\keyword{internal}
