% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{generate_table_sql_insert}
\alias{generate_table_sql_insert}
\title{Generate table sql insert}
\usage{
generate_table_sql_insert(table, instances)
}
\arguments{
\item{table}{A string, table name.}

\item{instances}{A \code{tibble}.}
}
\value{
A string.
}
\description{
Generate sql code for inserting a table.
}
\keyword{internal}
