% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_schema.R
\name{get_agg_functions}
\alias{get_agg_functions}
\title{Get aggregate functions}
\usage{
get_agg_functions(schema)
}
\arguments{
\item{schema}{A \code{fact_schema} object.}
}
\value{
A vector of strings.
}
\description{
Get aggregate functions
}
\keyword{internal}
