% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{get_all_dimension_operations}
\alias{get_all_dimension_operations}
\title{Gets the operations performed on a dimension in all \code{star_database} objects}
\usage{
get_all_dimension_operations(op_name, name, stars)
}
\arguments{
\item{op_name}{A string, operation name.}

\item{name}{A string, element name.}

\item{stars}{A list of \code{star_database} objects.}
}
\value{
A \code{star_operations} object.
}
\description{
Gets the operations performed on a dimension in all \code{star_database} objects
}
\keyword{internal}
