% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{get_attribute_names.star_schema}
\alias{get_attribute_names.star_schema}
\title{Get attribute names}
\usage{
\method{get_attribute_names}{star_schema}(schema)
}
\arguments{
\item{schema}{A \code{star_schema} object.}
}
\value{
A vector of strings.
}
\description{
Get the names of the attribute defined in the dimension schemas.
}
\keyword{internal}
