% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_schema.R
\name{get_attribute_names_schema.dimension_schema}
\alias{get_attribute_names_schema.dimension_schema}
\title{Get attribute names}
\usage{
\method{get_attribute_names_schema}{dimension_schema}(schema)
}
\arguments{
\item{schema}{A \code{dimension_schema} object.}
}
\value{
A string.
}
\description{
Get the attribute names.
}
\keyword{internal}
