% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_default_unknown_value}
\alias{get_default_unknown_value}
\title{get default unknown value}
\usage{
get_default_unknown_value()
}
\value{
A string.
}
\description{
get default unknown value
}
\keyword{internal}
