% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_schema.R
\name{get_dimension_name.dimension_schema}
\alias{get_dimension_name.dimension_schema}
\title{Get dimension name}
\usage{
\method{get_dimension_name}{dimension_schema}(schema)
}
\arguments{
\item{schema}{A \code{dimension_schema} object.}
}
\value{
A string.
}
\description{
Get the dimension name.
}
\keyword{internal}
