% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_dimension_names}
\alias{get_dimension_names}
\alias{get_dimension_names.star_database}
\title{Get the names of the dimensions of a star database}
\usage{
get_dimension_names(db, star)

\method{get_dimension_names}{star_database}(db, star = NULL)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{star}{A string or integer, star database name or index in constellation.}
}
\value{
A vector of strings, dimension names.
}
\description{
Obtain the names of the dimensions of a star database.
}
\examples{

names <- star_database(mrs_cause_schema, ft_num) |>
  get_dimension_names()

}
\seealso{
\code{\link{star_schema}}, \code{\link{flat_table}}

Other star database definition functions: 
\code{\link{get_dimension_table}()},
\code{\link{get_fact_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_table_names}()},
\code{\link{group_dimension_instances}()},
\code{\link{role_playing_dimension}()},
\code{\link{star_database}()}
}
\concept{star database definition functions}
