% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_dimension_table}
\alias{get_dimension_table}
\alias{get_dimension_table.star_database}
\title{Get dimension table}
\usage{
get_dimension_table(db, name)

\method{get_dimension_table}{star_database}(db, name = NULL)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}
}
\value{
A \code{tibble}, dimension table.
}
\description{
Get the table for the dimension indicated by its name.
}
\examples{

table <- star_database(mrs_cause_schema, ft_num) |>
  get_dimension_table("where")

}
\seealso{
\code{\link{star_schema}}, \code{\link{flat_table}}

Other star database definition functions: 
\code{\link{get_dimension_names}()},
\code{\link{get_fact_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_table_names}()},
\code{\link{group_dimension_instances}()},
\code{\link{role_playing_dimension}()},
\code{\link{star_database}()}
}
\concept{star database definition functions}
