% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{get_existing_fact_instances}
\alias{get_existing_fact_instances}
\alias{get_existing_fact_instances.star_database_update}
\title{Get existing fact instances}
\usage{
get_existing_fact_instances(sdbu)

\method{get_existing_fact_instances}{star_database_update}(sdbu)
}
\arguments{
\item{sdbu}{A \code{star_database_update} object.}
}
\value{
A \code{tibble} object.
}
\description{
From the planned update, it obtains the instances of the update facts that
are already included in the star database facts to be updated.
}
\details{
When the update occurs, we need to determine what happens to these instances.
}
\examples{

f1 <- flat_table('ft_num', ft_cause_rpd) |>
  as_star_database(mrs_cause_schema_rpd) |>
  replace_attribute_values(
    name = "When Available",
    old = c('1962', '11', '1962-03-14'),
    new = c('1962', '3', '1962-01-15')
  ) |>
  group_dimension_instances(name = "When")
f2 <- flat_table('ft_num2', ft_cause_rpd) |>
  update_according_to(f1)
fact_instances <- f2 |>
  get_existing_fact_instances()

}
\seealso{
Other star database refresh functions: 
\code{\link{get_new_dimension_instances}()},
\code{\link{update_according_to}()}
}
\concept{star database refresh functions}
