% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_schema.R
\name{get_fact_name.fact_schema}
\alias{get_fact_name.fact_schema}
\title{Get fact name}
\usage{
\method{get_fact_name}{fact_schema}(schema)
}
\arguments{
\item{schema}{A \code{fact_schema} object.}
}
\value{
A string.
}
\description{
Get fact name
}
\keyword{internal}
