% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_fact_names}
\alias{get_fact_names}
\alias{get_fact_names.star_database}
\title{Get the names of the facts of a star database}
\usage{
get_fact_names(db)

\method{get_fact_names}{star_database}(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A vector of strings, fact names.
}
\description{
Obtain the names of the facts of a star database.
}
\examples{

names <- star_database(mrs_cause_schema, ft_num) |>
  get_fact_names()

}
\seealso{
\code{\link{star_schema}}, \code{\link{flat_table}}

Other star database definition functions: 
\code{\link{get_dimension_names}()},
\code{\link{get_dimension_table}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_table_names}()},
\code{\link{group_dimension_instances}()},
\code{\link{role_playing_dimension}()},
\code{\link{star_database}()}
}
\concept{star database definition functions}
