% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_geoattribute_name}
\alias{get_geoattribute_name}
\title{Get geoattribute name}
\usage{
get_geoattribute_name(attribute)
}
\arguments{
\item{attribute}{A vector, attribute names.}
}
\value{
A string.
}
\description{
Get the name of the geoattribute from a vector of attribute names
}
\keyword{internal}
