% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolayer.R
\name{get_geolayer}
\alias{get_geolayer}
\alias{get_geolayer.geolayer}
\title{Get geographic layer}
\usage{
get_geolayer(gl)

\method{get_geolayer}{geolayer}(gl)
}
\arguments{
\item{gl}{A \code{geolayer} object.}
}
\value{
A \code{sf} object.
}
\description{
Get the geographic layer from a \code{geolayer} object.
}
\examples{

gl <- mrs_db_geo |>
  as_geolayer()

l <- gl |>
  get_geolayer()

}
\seealso{
Other query functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_geolayer}()},
\code{\link{filter_dimension}()},
\code{\link{get_variables}()},
\code{\link{run_query}()},
\code{\link{select_dimension}()},
\code{\link{select_fact}()},
\code{\link{set_variables}()},
\code{\link{star_query}()}
}
\concept{query functions}
