% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_layer_geometry}
\alias{get_layer_geometry}
\title{Get layer geometry}
\usage{
get_layer_geometry(layer)
}
\arguments{
\item{layer}{A \code{sf} object.}
}
\value{
A string.
}
\description{
Get the geometry of a layer. It will only be valid if one of the two geometries
is interpreted: \emph{point} or \emph{polygon}.
}
\examples{

geometry <- get_layer_geometry(us_layer_state)

}
\seealso{
Other star database geographic attributes: 
\code{\link{check_geoattribute_geometry}()},
\code{\link{define_geoattribute}()},
\code{\link{get_geoattribute_geometries}()},
\code{\link{get_geoattributes}()},
\code{\link{get_point_geometry}()},
\code{\link{summarize_layer}()}
}
\concept{star database geographic attributes}
