% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{get_measure_names_schema.fact_schema}
\alias{get_measure_names_schema.fact_schema}
\title{Get measure names}
\usage{
\method{get_measure_names_schema}{fact_schema}(schema)
}
\arguments{
\item{schema}{A \code{fact_schema} object.}
}
\value{
A vector of strings.
}
\description{
Get the names of the measures defined in the fact schema.
}
\keyword{internal}
