% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{get_measure_names_schema.star_schema}
\alias{get_measure_names_schema.star_schema}
\title{Get measure names}
\usage{
\method{get_measure_names_schema}{star_schema}(schema)
}
\arguments{
\item{schema}{A \code{star_schema} object.}
}
\value{
A vector of strings.
}
\description{
Get the names of the measures defined in the fact schema.
}
\keyword{internal}
