% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{get_new_dimension_instances}
\alias{get_new_dimension_instances}
\alias{get_new_dimension_instances.star_database_update}
\title{Get new dimension instances}
\usage{
get_new_dimension_instances(sdbu)

\method{get_new_dimension_instances}{star_database_update}(sdbu)
}
\arguments{
\item{sdbu}{A \code{star_database_update} object.}
}
\value{
A list of \code{tibble} objects.
}
\description{
From the planned update, it obtains the instances of the update dimensions
that are not included in the star database dimensions to be updated.
}
\examples{

f1 <- flat_table('ft_num', ft_cause_rpd) |>
  as_star_database(mrs_cause_schema_rpd) |>
  replace_attribute_values(
    name = "When Available",
    old = c('1962', '11', '1962-03-14'),
    new = c('1962', '3', '1962-01-15')
  ) |>
  group_dimension_instances(name = "When")
f2 <- flat_table('ft_num2', ft_cause_rpd) |>
  update_according_to(f1)
dim_instances <- f2 |>
  get_new_dimension_instances()

}
\seealso{
Other star database refresh functions: 
\code{\link{get_existing_fact_instances}()},
\code{\link{get_transformation_code}()},
\code{\link{get_transformation_file}()},
\code{\link{incremental_refresh}()},
\code{\link{update_according_to}()}
}
\concept{star database refresh functions}
