% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_operation.R
\name{get_next_operation.star_operation}
\alias{get_next_operation.star_operation}
\title{A \code{star_operation} object row is returned, the one following the actual given}
\usage{
\method{get_next_operation}{star_operation}(op, op_name, name = NULL, actual = NULL)
}
\arguments{
\item{op}{A \code{star_operation} object.}

\item{op_name}{A string, operation name.}

\item{name}{A string, element name.}

\item{actual}{A \code{star_operation} object.}
}
\value{
A data frame.
}
\description{
A \code{star_operation} object row is returned, the one following the actual given
}
\keyword{internal}
