% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_schema.R
\name{get_nrow_agg.fact_schema}
\alias{get_nrow_agg.fact_schema}
\title{Get number of rows aggregate column}
\usage{
\method{get_nrow_agg}{fact_schema}(schema)
}
\arguments{
\item{schema}{A \code{fact_schema} object.}
}
\value{
A string.
}
\description{
Get number of rows aggregate column
}
\keyword{internal}
