% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_schema.R
\name{get_nrow_agg}
\alias{get_nrow_agg}
\title{Get number of rows aggregate column}
\usage{
get_nrow_agg(schema)
}
\arguments{
\item{schema}{A \code{fact_schema} object.}
}
\value{
A string.
}
\description{
Get number of rows aggregate column
}
\keyword{internal}
