% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_point_geometry}
\alias{get_point_geometry}
\title{Get point geometry}
\usage{
get_point_geometry(layer)
}
\arguments{
\item{layer}{A \code{sf} object.}
}
\value{
A \code{sf} object.
}
\description{
Obtain point geometry from polygon geometry using the centroid.
}
\examples{

file <-
  system.file(
    "extdata",
    "layer_us_level.gpkg",
    package = "rolap"
  )
layer_us_state <- sf::st_read(file, 'layer_us_state',
                              quiet = TRUE)

layer <-
  get_point_geometry(layer_us_state)

}
\seealso{
Other star database geographic attributes: 
\code{\link{define_geoattribute}()},
\code{\link{get_geometry}()},
\code{\link{get_unrelated_instances}()},
\code{\link{summarize_layer}()}
}
\concept{star database geographic attributes}
