% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_purged_dimension_instances}
\alias{get_purged_dimension_instances}
\title{Get purged instances of dimensions}
\usage{
get_purged_dimension_instances(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A list of a \code{star_database} object and tables of deleted instances of
dimensions.
}
\description{
Get instances deleted from dimensions that are not referenced in the facts.
}
\keyword{internal}
