% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_playing.R
\name{get_rpd_dimensions}
\alias{get_rpd_dimensions}
\title{Get rpd dimensions of a dimension}
\usage{
get_rpd_dimensions(db, name)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}
}
\value{
A vector of dimension names.
}
\description{
Get rpd dimensions of a dimension
}
\keyword{internal}
