% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_similar_attribute_values}
\alias{get_similar_attribute_values}
\alias{get_similar_attribute_values.star_database}
\title{Get similar attribute values in a dimension}
\usage{
get_similar_attribute_values(db, name, attributes, column)

\method{get_similar_attribute_values}{star_database}(
  db,
  name,
  attributes = NULL,
  column = "dput_instance"
)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}

\item{attributes}{A vector of strings, attribute names.}

\item{column}{A string, name of the column to include a vector of values.}
}
\value{
A vector of 'tibble' objects with similar instances.
}
\description{
Get sets of attribute values in a dimension that differ only by tildes, spaces,
or punctuation marks.
}
\examples{

instances <- star_database(mrs_cause_schema, ft_num) |>
  get_similar_attribute_values(name = "where", attributes = c())

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_attribute_names}()},
\code{\link{get_measure_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{group_dimension_instances}()},
\code{\link{replace_attribute_values}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_attribute_names}()},
\code{\link{set_measure_names}()},
\code{\link{snake_case}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
