% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_similar_values_table}
\alias{get_similar_values_table}
\title{Get similar values in a table}
\usage{
get_similar_values_table(table, attributes, exclude_numbers, col_as_vector)
}
\arguments{
\item{table}{A \code{tibble} object.}

\item{attributes}{A vector of strings, attribute names.}

\item{exclude_numbers}{A boolean, exclude numbers from comparison.}

\item{col_as_vector}{A string, name of the column to include a vector of values.}
}
\value{
A vector of \code{tibble} objects with similar instances.
}
\description{
Get similar values in a table
}
\keyword{internal}
