% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{get_star_query_schema}
\alias{get_star_query_schema}
\title{Get star query schema}
\usage{
get_star_query_schema(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A star database schema, list of fact and dimension schemes.
}
\description{
Obtain the star database schema to perform queries.
}
\keyword{internal}
