% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{get_surrogate_key.dimension_table}
\alias{get_surrogate_key.dimension_table}
\title{Get surrogate key names}
\usage{
\method{get_surrogate_key}{dimension_table}(dimension_table)
}
\arguments{
\item{dimension_table}{A \code{dimension_table} object.}
}
\value{
A vector of strings.
}
\description{
Get the names of the surrogate keys defined in the dimension table.
}
\keyword{internal}
