% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{get_table}
\alias{get_table}
\alias{get_table.flat_table}
\title{Get the table of the flat table}
\usage{
get_table(ft)

\method{get_table}{flat_table}(ft)
}
\arguments{
\item{ft}{A \code{flat_table} object.}
}
\value{
A \code{tibble}, the table.
}
\description{
Obtain the table of a flat table.
}
\examples{

table <- flat_table('iris', iris) |>
  get_table()

}
\seealso{
\code{\link{flat_table}}

Other flat table definition functions: 
\code{\link{flat_table}()},
\code{\link{read_flat_table_file}()},
\code{\link{read_flat_table_folder}()},
\code{\link{select_attributes}()},
\code{\link{select_measures}()}
}
\concept{flat table definition functions}
