% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{get_unique_values_table}
\alias{get_unique_values_table}
\title{Get unique values in a table}
\usage{
get_unique_values_table(table, col_as_vector)
}
\arguments{
\item{table}{A \code{tibble} object.}

\item{col_as_vector}{A string, name of the column to include a vector of values.}
}
\value{
A vector of \code{tibble} objects with similar instances.
}
\description{
Get unique values in a table
}
\keyword{internal}
