% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{get_unknown_value_defined}
\alias{get_unknown_value_defined}
\alias{get_unknown_value_defined.flat_table}
\title{Get the unknown value defined}
\usage{
get_unknown_value_defined(ft)

\method{get_unknown_value_defined}{flat_table}(ft)
}
\arguments{
\item{ft}{A \code{flat_table} object.}
}
\value{
A string.
}
\description{
Obtain the unknown value of a flat table.
}
\examples{

table <- flat_table('iris', iris) |>
  get_unknown_value_defined()

}
\seealso{
\code{\link{select_attributes}}, \code{\link{select_measures}}

Other flat table definition functions: 
\code{\link{as_star_database}()},
\code{\link{flat_table}()},
\code{\link{get_table}()},
\code{\link{get_unknown_values}()},
\code{\link{read_flat_table_file}()},
\code{\link{read_flat_table_folder}()}
}
\concept{flat table definition functions}
