% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{group_facts}
\alias{group_facts}
\title{Group facts}
\usage{
group_facts(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\description{
Once the external keys have been possibly replaced, group the rows of facts.
}
\keyword{internal}
