% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{incremental_refresh}
\alias{incremental_refresh}
\alias{incremental_refresh.star_database}
\title{Refresh a star database in a constellation}
\usage{
incremental_refresh(db, sdbu, existing_instances, replace_transformations)

\method{incremental_refresh}{star_database}(
  db,
  sdbu,
  existing_instances = "ignore",
  replace_transformations = FALSE
)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{sdbu}{A \code{star_database_update} object.}

\item{existing_instances}{A string, operation to be carried out on the instances
of already existing facts. The possible values are: "ignore", "replace",
"group" and "delete".}

\item{replace_transformations}{A boolean, replace the \code{star_database}
transformation code with the \code{star_database_update} one.}
}
\value{
A \code{star_database} object.
}
\description{
Obtain the names of the tables of a star database.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num)
sdbu <- flat_table('ft_num2', ft_num) |>
  update_according_to(db)

db <- db |>
  incremental_refresh(sdbu)

}
\seealso{
Other star database refresh functions: 
\code{\link{get_existing_fact_instances}()},
\code{\link{get_new_dimension_instances}()},
\code{\link{get_transformation_code}()},
\code{\link{get_transformation_file}()},
\code{\link{update_according_to}()}
}
\concept{star database refresh functions}
