% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{integrate_geo_dimensions}
\alias{integrate_geo_dimensions}
\title{Integrate two geodimensions}
\usage{
integrate_geo_dimensions(gd1, gd2)
}
\arguments{
\item{gd1}{A geodimension.}

\item{gd2}{A geodimension.}
}
\value{
A geodimension.
}
\description{
Integrate two geodimensions
}
\keyword{internal}
