% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_add_custom_column}
\alias{interpret_operation_add_custom_column}
\title{Interpret operation}
\usage{
interpret_operation_add_custom_column(ft, op, file, last_op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{file}{file to write the code}

\item{last_op}{A boolean, is the last operation?}
}
\value{
A flat table.
}
\description{
operation,          name, details,                  details2
"add_custom_column", name, as.character(list(definition))
}
\details{
f <- function(...)
g <- as.character(list(f))
h <- eval(parse(text = g))
}
\keyword{internal}
