% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_group_dimension_instances}
\alias{interpret_operation_group_dimension_instances}
\title{Interpret operation}
\usage{
interpret_operation_group_dimension_instances(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,                  name
"group_dimension_instances", name)
}
\keyword{internal}
