% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_lookup_table}
\alias{interpret_operation_lookup_table}
\title{Interpret operation}
\usage{
interpret_operation_lookup_table(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,       name,        details,                           details2
"lookup_table", pk_attributes, c(attributes, '|', attribute_agg), c(measures, '|', measure_agg)
}
\keyword{internal}
