% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_replace_empty_values}
\alias{interpret_operation_replace_empty_values}
\title{Interpret operation}
\usage{
interpret_operation_replace_empty_values(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,             name,       details,                  details2
"replace_empty_values", attributes, empty_values
}
\keyword{internal}
