% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_role_playing_dimension}
\alias{interpret_operation_role_playing_dimension}
\title{Interpret operation}
\usage{
interpret_operation_role_playing_dimension(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,                name, details, details2
"role_playing_dimension",  rpd,  roles,   att_names)
}
\keyword{internal}
