% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_select_instances_by_comparison}
\alias{interpret_operation_select_instances_by_comparison}
\title{Interpret operation}
\usage{
interpret_operation_select_instances_by_comparison(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,                       name,              details,            details2
"select_instances_by_comparison", c(not, n_ele_set), unlist(attributes), c(unlist(comparisons), unlist(values))
}
\keyword{internal}
