% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_set_attribute_names}
\alias{interpret_operation_set_attribute_names}
\title{Interpret operation}
\usage{
interpret_operation_set_attribute_names(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,            name, details, details2
"set_attribute_names", name, old,     new)
}
\keyword{internal}
