% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_snake_case}
\alias{interpret_operation_snake_case}
\title{Interpret operation}
\usage{
interpret_operation_snake_case(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,
"snake_case")
}
\keyword{internal}
