% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_star_database}
\alias{interpret_operation_star_database}
\title{Interpret operation}
\usage{
interpret_operation_star_database(
  ft,
  op,
  schema,
  file,
  last_op,
  begin_in_star_database
)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{schema}{multidimensional schema}

\item{file}{file to write the code}

\item{last_op}{A boolean, is the last operation?}

\item{begin_in_star_database}{A boolean, is the first operation?}
}
\value{
A flat table.
}
\description{
operation,                name,    details,      details2
"star_database", names(db$schemas), unknown_value)
}
\keyword{internal}
