% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_transform_to_values}
\alias{interpret_operation_transform_to_values}
\title{Interpret operation}
\usage{
interpret_operation_transform_to_values(ft, op, file, last_op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{file}{file to write the code}

\item{last_op}{A boolean, is the last operation?}
}
\value{
A flat table.
}
\description{
operation,            name,      details, details2
"transform_to_values", attribute, measure, c(id_reverse, na_rm))
}
\keyword{internal}
