% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is_empty_string}
\alias{is_empty_string}
\title{check if a string is empty}
\usage{
is_empty_string(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A boolean.
}
\description{
check if a string is empty
}
\keyword{internal}
