% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_operation.R
\name{is_new_operation.star_operation}
\alias{is_new_operation.star_operation}
\title{A \code{star_operation} is new?}
\usage{
\method{is_new_operation}{star_operation}(op, op_name, name = NULL, details = NULL, details2 = NULL)
}
\arguments{
\item{op}{A \code{star_operation} object.}

\item{op_name}{A string, operation name.}

\item{name}{A string, element name.}

\item{details}{A vector of strings, operation details.}

\item{details2}{A vector of strings, operation additional details.}
}
\value{
A boolean.
}
\description{
A \code{star_operation} is new?
}
\keyword{internal}
