% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{is_scd}
\alias{is_scd}
\title{Is a scd dimension}
\usage{
is_scd(schema)
}
\arguments{
\item{schema}{A \code{dimension_schema} object.}
}
\value{
A boolean.
}
\description{
Is a scd dimension
}
\keyword{internal}
