% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{line_last_op}
\alias{line_last_op}
\title{Get line last operation}
\usage{
line_last_op(last_op)
}
\arguments{
\item{last_op}{A boolean, is the last operation?}
}
\value{
A string
}
\description{
Get line last operation
}
\keyword{internal}
