% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiple_value_key}
\alias{multiple_value_key}
\title{Multiple value key}
\usage{
multiple_value_key(tb, col_as_vector = NULL)
}
\arguments{
\item{tb}{A \code{tibble}.}

\item{col_as_vector}{A string, name of the column to include a vector of values.}
}
\value{
A \code{tibble}.
}
\description{
Gets the keys that have multiple values associated with them. The first field
in the table is the key, the rest of fields are the values.
}
\details{
If a name is indicated in the \code{col_as_vector} parameter, it includes a column
with the data in vector form to be used in other functions.
}
\examples{

tb <- unique(ft[, c('WEEK', 'Week Ending Date')])
mvk <- multiple_value_key(tb)

}
\concept{utility functions}
