% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolap.R
\name{olap_connection}
\alias{olap_connection}
\title{Specify connection parameters to connect to an MS OLAP Cube}
\usage{
olap_connection(
  userid = NULL,
  password = NULL,
  datasource,
  catalog,
  fileout = tempfile()
)
}
\arguments{
\item{userid}{Username or User ID. Defaults to saved credentials. If saved credentials are not available, Rstudio will ask for credentials.}

\item{password}{User password. Defaults to saved credentials. If saved credentials are not available, Rstudio will ask for credentials.}

\item{datasource}{A server address to the OLAP cube.}

\item{catalog}{The name of the cube.}

\item{fileout}{storage path for the intermediate csv file.}
}
\description{
This function returns a list formated with the correct parameters
needed to connect to an MS OLAP Cube using 'Rolap::read_olap'
}
\examples{
\dontrun{
con <- olap_connection("user", "pwd", "https://myolapcube.mycompany.com/msolap/", catalog = "Sales")
read_olap(
 con,
 "SELECT
    { [Measures].[Pack sales], [Measures].[Value sales] } ON COLUMNS,
    { [Year].[2020], [Year].[2021] } ON ROWS
  FROM [Sales]
  WHERE ( [Products].[Pharmaceuticals].[Oncology] )")
 }
}
\keyword{MDX}
\keyword{olap}
