% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_table.R
\name{prepare_instances_to_join}
\alias{prepare_instances_to_join}
\title{Transform the instances table implemented by a \code{tibble} to join}
\usage{
prepare_instances_to_join(instances)
}
\arguments{
\item{instances}{A \code{tibble}, the instances table.}
}
\value{
A \code{tibble}.
}
\description{
Transform all fields in the instances table to character type and replace
the \code{NA} values for a specific value.
}
\keyword{internal}
