% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{purge_dimension}
\alias{purge_dimension}
\title{Purge instances of a dimension}
\usage{
purge_dimension(db, dim)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{dim}{A string, dimension name.}
}
\value{
A \code{tibble}, dimension table.
}
\description{
Delete instances of a dimension that are not referenced in the facts.
}
\keyword{internal}
