% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{purge_dimension_instances}
\alias{purge_dimension_instances}
\title{Purge instances of dimensions}
\usage{
purge_dimension_instances(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A \code{star_database} object.
}
\description{
Delete instances of dimensions that are not referenced in the facts.
}
\keyword{internal}
