% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolap.R
\name{read_olap}
\alias{read_olap}
\title{Query an MS OLAP Cube}
\usage{
read_olap(con, query, clean_names = TRUE)
}
\arguments{
\item{con}{A list of connection parameters, as returned by 'Rolap::olap_connection'.}

\item{query}{A MDX-query.}

\item{clean_names}{Clean column names by the use of 'janitor::clean_names'?}
}
\description{
This function returns a tibble of data from an MDX queried OLAP cube.
}
\examples{
\dontrun{
con <- olap_connection("user", "pwd", "https://myolapcube.mycompany.com/msolap/", catalog = "Sales")
df <- read_olap(
 con,
 "SELECT
    { [Measures].[Pack sales], [Measures].[Value sales] } ON COLUMNS,
    { [Year].[2020], [Year].[2021] } ON ROWS
  FROM [Sales]
  WHERE ( [Products].[Pharmaceuticals].[Oncology] )")
 }
}
\keyword{MDX}
\keyword{olap}
