% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{reformat_file}
\alias{reformat_file}
\title{Get line last operation}
\usage{
reformat_file(out_file, function_name)
}
\arguments{
\item{out_file}{A string, file name.}

\item{function_name}{A string, name of the function to generate in the file.}
}
\value{
A string
}
\description{
Get line last operation
}
\keyword{internal}
