% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_new}
\alias{refresh_new}
\alias{refresh_new.flat_table}
\title{Refresh a new flat table}
\usage{
refresh_new(ft, s_op, star, sel_measure_group)

\method{refresh_new}{flat_table}(ft, s_op, star = 1, sel_measure_group = 1)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{s_op}{An object with defined modification operations.}

\item{star}{Star database name or index in constellation.}

\item{sel_measure_group}{A vector of integers, if measures are separated into
groups, indicate which group to consider.}
}
\value{
A \code{flat_table} object.
}
\description{
Refresh a new a flat table with the operations of another structure based on
a flat table.
}
\examples{

ft <- flat_table('iris', iris)

ft <- flat_table('ft_num', ft_num)

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{refresh}()},
\code{\link{remove_instances_without_measures}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{replace_unknown_values}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
