% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{remove_all_measures_na}
\alias{remove_all_measures_na}
\title{Delete if all measures are na}
\usage{
remove_all_measures_na(table, measures)
}
\arguments{
\item{table}{A \code{tibble} object.}

\item{measures}{A vector of strings, measure names.}
}
\description{
Delete if all measures are na
}
\keyword{internal}
