% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{remove_instances_without_measures}
\alias{remove_instances_without_measures}
\alias{remove_instances_without_measures.flat_table}
\title{Remove instances without measures}
\usage{
remove_instances_without_measures(ft)

\method{remove_instances_without_measures}{flat_table}(ft)
}
\arguments{
\item{ft}{A \code{flat_table} object.}
}
\value{
A \code{flat_table} object.
}
\description{
Delete instances that have all measures undefined.
}
\examples{

ft <- flat_table('iris', iris) |>
  remove_instances_without_measures()

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{replace_unknown_values}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
