% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{replace_empty_values}
\alias{replace_empty_values}
\alias{replace_empty_values.flat_table}
\title{Replace empty values with the unknown value}
\usage{
replace_empty_values(ft, attributes)

\method{replace_empty_values}{flat_table}(ft, attributes = NULL)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{attributes}{A vector of names.}
}
\value{
A \code{flat_table} object.
}
\description{
Transforms the given attributes by replacing the empty values with the unknown
value.
}
\examples{

ft <- flat_table('iris', iris) |>
  replace_empty_values()

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{replace_string}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
