% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{replace_empty_values_table}
\alias{replace_empty_values_table}
\title{Replace empty values with the unknown value}
\usage{
replace_empty_values_table(
  table,
  attributes = NULL,
  empty_values = NULL,
  unknown_value
)
}
\arguments{
\item{table}{A \code{tibble} object.}

\item{attributes}{A vector of names.}

\item{empty_values}{A vector of values that correspond to empty values.}

\item{unknown_value}{A string.}
}
\value{
A \code{tibble} object.
}
\description{
Replace empty values with the unknown value
}
\keyword{internal}
