% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{replace_instance_values}
\alias{replace_instance_values}
\alias{replace_instance_values.star_database}
\title{Replace instance values in a dimension}
\usage{
replace_instance_values(db, name, old, new)

\method{replace_instance_values}{star_database}(db, name, old, new)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}

\item{old}{A vector of values.}

\item{new}{A vector of values.}
}
\value{
A \code{star_database} object.
}
\description{
Given the values of a possible instance of a dimension, for that combination,
replace them with the new data values.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num) |>
  replace_instance_values(name = "where",
    old = c('1', 'CT', 'Bridgeport'),
    new = c('1', 'CT', 'Hartford'))

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_dimension_attribute_names}()},
\code{\link{get_fact_measure_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_similar_instances}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_dimension_attribute_names}()},
\code{\link{set_fact_measure_names}()},
\code{\link{snake_case}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
