% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{replace_names}
\alias{replace_names}
\title{Replace names}
\usage{
replace_names(names, old, new)
}
\arguments{
\item{names}{A string, original names.}

\item{old}{A vector of names to replace.}

\item{new}{A vector of names, new names.}
}
\value{
A vector of strings, names replaced.
}
\description{
Replace names
}
\keyword{internal}
