% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{replace_unknown_values}
\alias{replace_unknown_values}
\alias{replace_unknown_values.flat_table}
\title{Replace unknown values with the given value}
\usage{
replace_unknown_values(ft, attributes, value)

\method{replace_unknown_values}{flat_table}(ft, attributes = NULL, value)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{attributes}{A vector of names.}

\item{value}{A value.}
}
\value{
A \code{flat_table} object.
}
\description{
Transforms the given attributes by replacing unknown values in them with
the given value.
}
\examples{

iris2 <- iris
iris2[10, 'Species'] <- NA
ft <- flat_table('iris', iris2) |>
  replace_empty_values() |>
  replace_unknown_values(value = "Not available")

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{remove_instances_without_measures}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
