% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{rpd_in_constellation}
\alias{rpd_in_constellation}
\title{Transform role playing dimensions in constellation}
\usage{
rpd_in_constellation(db)
}
\arguments{
\item{db}{A \code{constellation} object.}
}
\value{
A \code{constellation} object.
}
\description{
Transform role playing dimensions in constellation
}
\keyword{internal}
