% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolayer.R
\name{same_granularity_facts}
\alias{same_granularity_facts}
\title{Do all fact tables have the same granularity?}
\usage{
same_granularity_facts(db, names)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{names}{A vector of strings, fact names.}
}
\value{
A boolean.
}
\description{
Do all fact tables have the same granularity?
}
\keyword{internal}
