% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{select_attributes}
\alias{select_attributes}
\alias{select_attributes.flat_table}
\title{Select attributes of a flat table}
\usage{
select_attributes(ft, attributes)

\method{select_attributes}{flat_table}(ft, attributes)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{attributes}{A vector of names.}
}
\value{
A \code{flat_table} object.
}
\description{
Select only the indicated attributes from the flat table.
}
\examples{

ft <- flat_table('iris', iris) |>
  select_attributes(attributes = c('Species'))

ft <- flat_table('ft_num', ft_num) |>
  select_attributes(attributes = c('Year', 'WEEK', 'Week Ending Date'))

}
\seealso{
\code{\link{flat_table}}

Other flat table definition functions: 
\code{\link{flat_table}()},
\code{\link{get_table}()},
\code{\link{read_flat_table_file}()},
\code{\link{read_flat_table_folder}()},
\code{\link{select_measures}()}
}
\concept{flat table definition functions}
