% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{select_measure_group}
\alias{select_measure_group}
\title{Select a measure group from a list}
\usage{
select_measure_group(lft, name)
}
\arguments{
\item{lft}{A list of \code{flat_table} objects.}

\item{name}{A string, measure group name.}
}
\value{
A \code{flat_table} object.
}
\description{
Once the measures are separated into groups, we have a list of flat tables,
we select one of the groups using this function, indicating its name.
}
\details{
The operation is registered, to know which group we will continue working
with. For this reason we should not use direct selection of elements from a
list.
}
\examples{

lft <- flat_table('iris', iris) |>
  separate_measures(
    measures = list(
      c('Petal.Length'),
      c('Petal.Width'),
      c('Sepal.Length'),
      c('Sepal.Width')
    ),
    names = c('PL', 'PW', 'SL', 'SW')
  )
ft <- lft |>
  select_measure_group('SL')

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{remove_instances_without_measures}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{replace_unknown_values}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()},
\code{\link{update_according_to}()}
}
\concept{flat table transformation functions}
