% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_transform.R
\name{separate_measures}
\alias{separate_measures}
\alias{separate_measures.flat_table}
\title{Separate measures in flat tables}
\usage{
separate_measures(ft, measures, names)

\method{separate_measures}{flat_table}(ft, measures = NULL, names = NULL)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{measures}{A list of string vectors, groups of measure names.}

\item{names}{A list of string, measure group names.}
}
\value{
A list of \code{flat_table} objects.
}
\description{
Separate measures listed as list items into flat tables. Each item in the
list is a vector of measures that is uniquely included along with the
attributes in a new flat table.
}
\details{
A list of flat tables is returned. It assign the names to the result list.
}
\examples{

lft <- flat_table('iris', iris) |>
  separate_measures(
    measures = list(
      c('Petal.Length'),
      c('Petal.Width'),
      c('Sepal.Length'),
      c('Sepal.Width')
    ),
    names = c('PL', 'PW', 'SL', 'SW')
  )

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{select_attributes}()},
\code{\link{select_measures}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
