% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{share_dimension_instance_operations}
\alias{share_dimension_instance_operations}
\title{Share dimension instance operations between all \code{star_database} objects}
\usage{
share_dimension_instance_operations(stars, dim_freq)
}
\arguments{
\item{stars}{A list of \code{star_database} objects.}

\item{dim_freq}{Dimension frequency table.}
}
\value{
A list of \code{star_database} objects.
}
\description{
Share dimension instance operations between all \code{star_database} objects
}
\keyword{internal}
