% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_playing.R
\name{simplify_rpd_dimensions}
\alias{simplify_rpd_dimensions}
\title{From a vector of dimensions, leave only one of each rpd.}
\usage{
simplify_rpd_dimensions(db, names)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{names}{A vector of strings, dimension names.}
}
\value{
A vector of dimension names.
}
\description{
From a vector of dimensions, leave only one of each rpd.
}
\keyword{internal}
