% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table.R
\name{snake_case_table.fact_table}
\alias{snake_case_table.fact_table}
\title{Transform names according to the snake case style}
\usage{
\method{snake_case_table}{fact_table}(table)
}
\arguments{
\item{table}{A \code{fact_table} object.}
}
\value{
A \code{fact_table} object.
}
\description{
Transform names according to the snake case style
}
\keyword{internal}
