% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{star_database_with_previous_operations}
\alias{star_database_with_previous_operations}
\title{Creates a \code{star_database} adding previous operations}
\usage{
star_database_with_previous_operations(
  schema,
  instances,
  unknown_value = NULL,
  operations = NULL
)
}
\arguments{
\item{schema}{A \code{star_schema} object.}

\item{instances}{A flat table to define the database instances according to the schema.}

\item{unknown_value}{A string, value used to replace NA values in dimensions.}

\item{operations}{A list of operations.}
}
\value{
A \code{star_database} object.
}
\description{
Creates a \code{star_database} adding previous operations
}
\keyword{internal}
