% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{star_operation}
\alias{star_operation}
\title{\code{star_operation} S3 class}
\usage{
star_operation()
}
\value{
A \code{star_operation} object.
}
\description{
A \code{star_operation} object is created.
}
\details{
A \code{star_operation} object is part of a \code{star_schema} object, defines
operations of the star schema.
}
\keyword{internal}
