% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_query.R
\name{star_query}
\alias{star_query}
\alias{star_query.star_database}
\title{\code{star_query} S3 class}
\usage{
star_query(db)

\method{star_query}{star_database}(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A \code{star_query} object.
}
\description{
An empty \code{star_query} object is created where we can select facts and
measures, dimensions, dimension attributes and filter dimension rows.
}
\examples{

sq <- mrs_db |>
  star_query()

}
\seealso{
Other query functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_geolayer}()},
\code{\link{filter_dimension}()},
\code{\link{get_layer}()},
\code{\link{get_variable_description}()},
\code{\link{get_variables}()},
\code{\link{run_query}()},
\code{\link{select_dimension}()},
\code{\link{select_fact}()},
\code{\link{set_layer}()},
\code{\link{set_variables}()}
}
\concept{query functions}
