% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{star_schema}
\alias{star_schema}
\title{\code{star_schema} S3 class}
\usage{
star_schema()
}
\value{
A \code{star_schema} object.
}
\description{
An empty \code{star_schema} object is created in which definition of facts
and dimensions can be added.
}
\details{
To get a star database (a \code{star_database} object) we need a flat table
(implemented through a \code{tibble}) and a \code{star_schema} object. The
definition of facts and dimensions in the \code{star_schema} object is made
from the flat table columns. Each attribute can only appear once in the
definition.
}
\examples{

sc <- star_schema()

}
\seealso{
\code{\link{star_schema}}
}
\concept{star definition functions}
