% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{string_or_null}
\alias{string_or_null}
\title{Get the representation to output}
\usage{
string_or_null(value)
}
\arguments{
\item{value}{A string}
}
\value{
A string
}
\description{
Get the representation to output
}
\keyword{internal}
