% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{summarize_layer}
\alias{summarize_layer}
\title{Summarize geometry of a layer}
\usage{
summarize_layer(from_layer, attribute)
}
\arguments{
\item{from_layer}{A \code{sf} object.}

\item{attribute}{A string, attribute name.}
}
\value{
A \code{sf} object.
}
\description{
Groups the geometric elements of a layer according to the values of the indicated
attribute.
}
\keyword{internal}
