% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{transform_names}
\alias{transform_names}
\title{For each row, add a vector of values}
\usage{
transform_names(names, ordered, as_definition)
}
\arguments{
\item{names}{A vector of strings, names of attributes or measures.}

\item{ordered}{A boolean, sort names alphabetically.}

\item{as_definition}{A boolean, as the definition of the vector in R.}
}
\value{
A vector of strings, attribute or measure names.
}
\description{
For each row, add a vector of values
}
\keyword{internal}
