% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolayer.R
\name{unify_facts_and_dimensions}
\alias{unify_facts_and_dimensions}
\title{Unify facts and dimensions in a flat table}
\usage{
unify_facts_and_dimensions(db, dimension, include_nrow_agg)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{dimension}{A vector of strings, dimension names.}

\item{include_nrow_agg}{A boolean.}
}
\value{
A \code{tibble}.
}
\description{
Unify facts and dimensions in a flat table
}
\keyword{internal}
