% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{unify_rpd}
\alias{unify_rpd}
\title{Unify lists of dimension names if there are any in common}
\usage{
unify_rpd(rpd)
}
\arguments{
\item{rpd}{A list of strings (dimension names).}
}
\description{
Unify lists of dimension names if there are any in common
}
\keyword{internal}
