% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_layer_state}
\alias{us_layer_state}
\title{Geographic layer of US States}
\format{
A \code{sf}.
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-data.2021.html}
}
\usage{
us_layer_state
}
\description{
Geographic layer with data from the States of the USA in polygon format, with
simplified geometry so that it takes up less space.
}
\details{
It has been obtained from the geographic data included in the US census prepared
by the U.S. Census Bureau.
}
\keyword{datasets}
