% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{validate_dimension_attributes}
\alias{validate_dimension_attributes}
\title{Validate dimension attributes}
\usage{
validate_dimension_attributes(db, dimension, attributes)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{dimension}{A dimension name.}

\item{attributes}{Attribute names.}
}
\value{
A vector of strings, dimension names.
}
\description{
Validate dimension attributes
}
\keyword{internal}
