% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{validate_dimension_names}
\alias{validate_dimension_names}
\title{Validate dimension names}
\usage{
validate_dimension_names(db, name)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A vector of strings, dimension names.}
}
\value{
A vector of strings, dimension names.
}
\description{
Validate dimension names
}
\keyword{internal}
