% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{validate_facts}
\alias{validate_facts}
\title{Validate fact names}
\usage{
validate_facts(defined_facts, facts)
}
\arguments{
\item{defined_facts}{A vector of strings, defined fact names.}

\item{measures}{A vector of strings, fact names.}
}
\value{
A vector of strings, fact names.
}
\description{
Validate fact names
}
\keyword{internal}
