% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table_join.R
\name{validate_lookup_parameters}
\alias{validate_lookup_parameters}
\title{Validate lookup parameters}
\usage{
validate_lookup_parameters(ft, fk_attributes, lookup)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{fk_attributes}{A vector of strings, attribute names.}

\item{lookup}{A \code{flat_table} object.}
}
\value{
A vector of strings, fk attribute names.
}
\description{
Validate lookup parameters
}
\keyword{internal}
