% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{validate_measures}
\alias{validate_measures}
\title{Validate measure names}
\usage{
validate_measures(defined_measures, measures)
}
\arguments{
\item{defined_measures}{A vector of strings, defined measure names.}

\item{measures}{A vector of strings, measure names.}
}
\value{
A vector of strings, measure names.
}
\description{
Validate measure names
}
\keyword{internal}
