% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{vector_presentation}
\alias{vector_presentation}
\title{vector to string for presentation}
\usage{
vector_presentation(vector)
}
\arguments{
\item{vector}{A vector}
}
\value{
A string
}
\description{
vector to string for presentation
}
\keyword{internal}
