% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{vector_to_string}
\alias{vector_to_string}
\title{Transforms a vector of strings into a string.}
\usage{
vector_to_string(vector)
}
\arguments{
\item{vector}{A vector of strings.}
}
\value{
A string.
}
\description{
Transforms a vector of strings into a string.
}
\keyword{internal}
