% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{check_refesh}
\alias{check_refesh}
\title{Checks the refresh of the selected star database from the given database}
\usage{
check_refesh(db, refresh_db)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{refresh_db}{A \code{star_database} object with the same structure with
new data.}
}
\value{
A list of facts and dimensions, first facts, then dimensions.
}
\description{
Checks the refresh operation of the selected star database from the given
database. Once this operation is carried out, the results can be consulted on
the new instances in dimensions or existing instances in the facts.
}
\keyword{internal}
