% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{constellation}
\alias{constellation}
\title{\code{constellation} S3 class}
\usage{
constellation(name = NULL, stars = NULL)
}
\arguments{
\item{name}{A string.}

\item{stars}{A list of \code{star_database} objects.}
}
\value{
A \code{constellation} object.
}
\description{
Creates a \code{constellation} object from a list of \code{star_database} objects. All
dimensions with the same name in the star schemas have to be conformable
(share the same structure, even though they have different instances).
}
\examples{

when <- dimension_schema(name = "When",
                         attributes = c("Year"))

where <- dimension_schema(name = "Where",
                          attributes = c("REGION",
                                         "State",
                                         "City"))

s1 <- star_schema() |>
  define_facts(name = "MRS Cause",
               measures = c("Pneumonia and Influenza Deaths",
                            "All Deaths")) |>
  define_dimension(when) |>
  define_dimension(where)

db1 <- star_database(s1, ft_num) |>
  snake_case()


s2 <- star_schema() |>
  define_facts(name = "MRS Age",
               measures = c("All Deaths")) |>
  define_dimension(when) |>
  define_dimension(where) |>
  define_dimension(name = "Who",
                   attributes = c("Age"))

db2 <- star_database(s2, ft_age) |>
  snake_case()

ct <- constellation("MRS", list(db1, db2))

}
\seealso{
Other star schema and constellation definition functions: 
\code{\link{snake_case}()}
}
\concept{star schema and constellation definition functions}
