% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_finest}
\alias{db_finest}
\title{Czech debit card company specialising on payments at gas stations (finest detail)}
\format{
A \code{star_database}.
}
\source{
\url{https://fit.cvut.cz/cs}
}
\usage{
db_finest
}
\description{
Multidimensional design with finest detail from the data available at the source.
}
\seealso{
Other debit card example data: 
\code{\link{db_summary}}
}
\concept{debit card example data}
\keyword{datasets}
