% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_summary}
\alias{db_summary}
\title{Czech debit card company specialising on payments at gas stations (summary)}
\format{
A \code{star_database}.
}
\source{
\url{https://relational.fit.cvut.cz/dataset/CCS}
}
\usage{
db_summary
}
\description{
Multidimensional design with a summary from the data available at the source.
}
\details{
The development has been done in the vignette \code{vignette("rdbms-dm")}.
}
\seealso{
\code{\link{mrs_cause_schema}}
}
\concept{database example data}
\keyword{datasets}
