% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{define_dimension}
\alias{define_dimension}
\title{Define dimension in a \code{star_schema} object.}
\usage{
define_dimension(star, dimension)
}
\arguments{
\item{star}{A \code{star_schema} object.}

\item{dimension}{A \code{dimension_schema} object.}
}
\value{
A \code{star_schema} object.
}
\description{
Define dimension in a \code{star_schema} object.
}
\examples{

s <- star_schema()
d <- dimension_schema(
  name = "when",
  attributes = c(
    "Week Ending Date",
    "WEEK",
    "Year"
  )
)
s <- s |>
  define_dimension(d)
}
\seealso{
\code{\link{star_schema}}

Other star schema definition functions: 
\code{\link{define_facts}()},
\code{\link{dimension_schema}()},
\code{\link{fact_schema}()},
\code{\link{star_schema}()}
}
\concept{star schema definition functions}
