% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_operation_set}
\alias{delete_operation_set}
\title{Delete a set of operations}
\usage{
delete_operation_set(op, op2)
}
\arguments{
\item{op}{A \code{star_operation} object.}

\item{op2}{A \code{star_operation} object.}
}
\value{
op A \code{star_operation} object.
}
\description{
Delete a set of operations
}
\keyword{internal}
