% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\alias{deploy.star_database}
\title{Deploy a star database in a relational database}
\usage{
deploy(db, name, connect, disconnect, file)

\method{deploy}{star_database}(db, name, connect, disconnect = NULL, file = NULL)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, name of the deployment.}

\item{connect}{A function that returns a \code{DBI::DBIConnection} object.}

\item{disconnect}{A function that receives a \code{DBI::DBIConnection} object as a
parameter and close the connection.}

\item{file}{A string, name of the file to store the object.}
}
\value{
A \code{star_database} object.
}
\description{
Deploy a star database in a relational database
}
\examples{

f1 <-
  flat_table('ft_num', ft_cause_rpd[ft_cause_rpd$City != 'Cambridge' &
                                      ft_cause_rpd$WEEK != '4',]) |>
  as_star_database(mrs_cause_schema_rpd) |>
  role_playing_dimension(rpd = "When",
                         roles = c("When Available", "When Received"))
f2 <- flat_table('ft_num2', ft_cause_rpd[ft_cause_rpd$City != 'Bridgeport' &
                                           ft_cause_rpd$WEEK != '2',])
f2 <- f2 |>
  update_according_to(f1)

f1 <- f1 |>
  incremental_refresh(f2)

}
\seealso{
Other star database refresh functions: 
\code{\link{get_existing_fact_instances}()},
\code{\link{get_lookup_tables}()},
\code{\link{get_new_dimension_instances}()},
\code{\link{get_star_database}()},
\code{\link{get_star_schema}()},
\code{\link{get_transformation_code}()},
\code{\link{get_transformation_file}()},
\code{\link{incremental_refresh}()},
\code{\link{update_according_to}()}
}
\concept{star database refresh functions}
