% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{dimension_table}
\alias{dimension_table}
\title{\code{dimension_table} S3 class}
\usage{
dimension_table(name = NULL, attributes = NULL, instances = NULL)
}
\arguments{
\item{name}{A string, dimension name.}

\item{attributes}{A vector of strings, attributes names.}

\item{instances}{A flat table with the dimension instances.}
}
\value{
A \code{dimension_table} object.
}
\description{
A \code{dimension_table} object is created, we have to define its
surrogate key.
}
\keyword{internal}
