% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_table.R
\name{fact_table}
\alias{fact_table}
\title{\code{fact_table} S3 class}
\usage{
fact_table(
  name = NULL,
  surrogate_keys = NULL,
  dim_int_names = NULL,
  instances = NULL
)
}
\arguments{
\item{name}{A string, fact name.}

\item{surrogate_keys}{A vector of strings, surrogate key names.}

\item{dim_int_names}{A vector of strings, internal names of dimensions.}

\item{instances}{A flat table with the fact instances.}
}
\value{
A \code{fact_table} object.
}
\description{
A \code{fact_table} object is created, we have to get its
surrogate keys.
}
\keyword{internal}
