% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft}
\alias{ft}
\title{Mortality Reporting System}
\format{
A \code{tibble}.
}
\source{
\url{https://catalog.data.gov/dataset/deaths-in-122-u-s-cities-1962-2016-122-cities-mortality-reporting-system}
}
\usage{
ft
}
\description{
Selection of 20 rows from the 122 Cities Mortality Reporting System.
}
\details{
The original dataset covers from 1962 to 2016. For each week, in 122 US cities,
mortality figures by age group and cause, considered separately, are included.
In the cause, only a distinction is made between pneumonia or influenza and
others.
}
\seealso{
\code{\link{mrs_cause_schema}}

Other mrs example data: 
\code{\link{ft_age}},
\code{\link{ft_num}}
}
\concept{mrs example data}
\keyword{datasets}
