% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_geometry}
\alias{get_geometry}
\title{Get geometry}
\usage{
get_geometry(layer)
}
\arguments{
\item{layer}{A \code{sf} object.}
}
\value{
A string.
}
\description{
Get the geometry of a layer, as it is interpreted to define a \code{geolevel}
object.
}
\details{
It will only be valid if one of the two geometries is interpreted: \emph{point}
or \emph{polygon}.
}
\examples{

geometry <- get_geometry(us_layer_state)

}
\seealso{
Other star database geographic attributes: 
\code{\link{define_geoattribute}()},
\code{\link{get_point_geometry}()},
\code{\link{get_unrelated_instances}()},
\code{\link{summarize_layer}()}
}
\concept{star database geographic attributes}
