% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_point_geometry}
\alias{get_point_geometry}
\title{Get point geometry}
\usage{
get_point_geometry(layer)
}
\arguments{
\item{layer}{A \code{sf} object.}
}
\value{
A \code{sf} object.
}
\description{
Obtain point geometry from polygon geometry.
}
\examples{

layer <-
  get_point_geometry(us_layer_state)

}
\seealso{
Other star database geographic attributes: 
\code{\link{check_geoattribute_geometry}()},
\code{\link{define_geoattribute}()},
\code{\link{get_geoattribute_geometries}()},
\code{\link{get_geoattributes}()},
\code{\link{get_layer_geometry}()},
\code{\link{summarize_layer}()}
}
\concept{star database geographic attributes}
