% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{get_role_playing_dimension_names}
\alias{get_role_playing_dimension_names}
\alias{get_role_playing_dimension_names.star_database}
\title{Get the names of the role playing dimensions}
\usage{
get_role_playing_dimension_names(db)

\method{get_role_playing_dimension_names}{star_database}(db)
}
\arguments{
\item{db}{A \code{constellation} object.}
}
\value{
A list of vector of strings with dimension names.
}
\description{
Role playing dimensions are defined in star_databases. When integrating
several star_databases to form a constellation, role playing dimensions are
also integrated. This function allows you to see the result.
}
\examples{

db1 <- star_database(mrs_cause_schema_rpd, ft_cause_rpd) |>
  role_playing_dimension(
    rpd = "When",
    roles = c("When Available", "When Received")
  )

db2 <- star_database(mrs_age_schema_rpd, ft_age_rpd) |>
  role_playing_dimension(
    rpd = "When Arrived",
    roles = c("When Available")
  )
rpd <- constellation("MRS", list(db1, db2)) |>
  get_role_playing_dimension_names()

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_attribute_names}()},
\code{\link{get_dimension_names}()},
\code{\link{get_measure_names}()},
\code{\link{get_similar_attribute_values_individually}()},
\code{\link{get_similar_attribute_values}()},
\code{\link{get_table_names}()},
\code{\link{get_unique_attribute_values}()},
\code{\link{group_dimension_instances}()},
\code{\link{replace_attribute_values}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_attribute_names}()},
\code{\link{set_measure_names}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
