% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_similar_attribute_values}
\alias{get_similar_attribute_values}
\alias{get_similar_attribute_values.star_database}
\title{Get similar attribute values combination in dimensions}
\usage{
get_similar_attribute_values(db, name, attributes, col_as_vector)

\method{get_similar_attribute_values}{star_database}(
  db,
  name = NULL,
  attributes = NULL,
  col_as_vector = NULL
)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}

\item{attributes}{A vector of strings, attribute names.}

\item{col_as_vector}{A string, name of the column to include a vector of values.}
}
\value{
A vector of \code{tibble} objects with similar instances.
}
\description{
Get sets of attribute values in dimensions that differ only by tildes, spaces,
or punctuation marks, for the combination of the given set of attributes.
}
\details{
A list of dimensions can be indicated, otherwise it considers all dimensions.

If a dimension is indicated, a list of attributes to be considered in it can
also be indicated. If several dimensions are indicated, the combination of
all the attributes of each dimension is considered.

If a name is indicated in the \code{col_as_vector} parameter, it includes a column
with the data in vector form to be used in other functions.
}
\examples{

instances <- star_database(mrs_cause_schema, ft_num) |>
  get_similar_attribute_values(name = "where")

db <- star_database(mrs_cause_schema, ft_num)
db$dimensions$where$table$City[2] <- " BrId  gEport "
db |>
  get_similar_attribute_values("where")

db <- star_database(mrs_cause_schema, ft_num)
db$dimensions$where$table$City[2] <- " BrId  gEport "
db |>
  get_similar_attribute_values("where",
    attributes = c("City", "State"),
    col_as_vector = "As a vector")

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_attribute_names}()},
\code{\link{get_dimension_names}()},
\code{\link{get_measure_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_similar_attribute_values_individually}()},
\code{\link{get_table_names}()},
\code{\link{get_unique_attribute_values}()},
\code{\link{group_dimension_instances}()},
\code{\link{replace_attribute_values}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_attribute_names}()},
\code{\link{set_measure_names}()},
\code{\link{snake_case}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
