% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{get_similar_attribute_values_individually}
\alias{get_similar_attribute_values_individually}
\alias{get_similar_attribute_values_individually.star_database}
\title{Get similar attribute values for individual attributes in dimensions}
\usage{
get_similar_attribute_values_individually(
  db,
  name,
  exclude_numbers,
  col_as_vector
)

\method{get_similar_attribute_values_individually}{star_database}(
  db,
  name = NULL,
  exclude_numbers = FALSE,
  col_as_vector = NULL
)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A vector of strings, dimension names.}

\item{exclude_numbers}{A boolean, exclude numbers from comparison.}

\item{col_as_vector}{A string, name of the column to include a vector of values.}
}
\value{
A vector of \code{tibble} objects with similar instances.
}
\description{
Get sets of attribute values for individual attributes in the set of dimensions
that differ only by tildes, spaces, or punctuation marks. If no dimension name
is indicated, all dimensions are considered.
}
\details{
You can indicate that the numbers are ignored to make the comparison.

If a name is indicated in the \code{col_as_vector} parameter, it includes a column
with the data in vector form to be used in other functions.
}
\examples{

instances <- star_database(mrs_cause_schema, ft_num) |>
  get_similar_attribute_values_individually(name = c("where", "when"))

instances <- star_database(mrs_cause_schema, ft_num) |>
  get_similar_attribute_values_individually()

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_attribute_names}()},
\code{\link{get_dimension_names}()},
\code{\link{get_measure_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_similar_attribute_values}()},
\code{\link{get_table_names}()},
\code{\link{get_unique_attribute_values}()},
\code{\link{group_dimension_instances}()},
\code{\link{replace_attribute_values}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_attribute_names}()},
\code{\link{set_measure_names}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
