% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{get_unrelated_instances}
\alias{get_unrelated_instances}
\alias{get_unrelated_instances.star_database}
\title{Get unrelated instances of a \code{geoattribute}}
\usage{
get_unrelated_instances(db, dimension, attribute, geometry)

\method{get_unrelated_instances}{star_database}(
  db,
  dimension = NULL,
  attribute = NULL,
  geometry = "polygon"
)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{dimension}{A string, dimension name.}

\item{attribute}{A vector, attribute names.}

\item{geometry}{A string, geometry name ('point' or 'polygon').}
}
\value{
A \code{tibble}.
}
\description{
We obtain the values of the dimension attribute that do not have an associated
geographic element of the indicated geometry.
}
\examples{

db <- mrs_db |>
  define_geoattribute(
    dimension = "where",
    attribute = "state",
    from_layer = us_layer_state,
    by = "STUSPS"
  )

instances <- get_unrelated_instances (db,
                                      dimension = "where",
                                      attribute = "state")

}
\seealso{
Other star database geographic attributes: 
\code{\link{define_geoattribute}()},
\code{\link{get_geometry}()},
\code{\link{get_point_geometry}()},
\code{\link{summarize_layer}()}
}
\concept{star database geographic attributes}
