% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_table.R
\name{group_by_keys}
\alias{group_by_keys}
\title{Group instances by keys aggregating the measures using the corresponding
aggregation function.}
\usage{
group_by_keys(instances, keys, measures, agg_functions, nrow_agg)
}
\arguments{
\item{instances}{A \code{tibble}, the instances table.}

\item{keys}{A vector of strings, key names to group by.}

\item{measures}{A vector of strings, measures to aggregate.}

\item{agg_functions}{A vector of strings, aggregate functions.}

\item{nrow_agg}{A string, name of a new column to count the number of rows
aggregated.}
}
\value{
A \code{tibble}.
}
\description{
Group instances by keys aggregating the measures using the corresponding
aggregation function.
}
\keyword{internal}
