% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{group_dimension_instances}
\alias{group_dimension_instances}
\alias{group_dimension_instances.star_database}
\title{Group instances of a dimension}
\usage{
group_dimension_instances(db, name)

\method{group_dimension_instances}{star_database}(db, name)
}
\arguments{
\item{db}{A \code{star_database} object.}

\item{name}{A string, dimension name.}
}
\value{
A \code{star_database} object.
}
\description{
After changes in values in the instances of a dimension, groups the instances
and, if necessary, also the related facts.
}
\examples{

db <- star_database(mrs_cause_schema, ft_num) |>
  group_dimension_instances(name = "where")

}
\seealso{
\code{\link{as_tibble_list}}, \code{\link{as_dm_class}}

Other star database and constellation definition functions: 
\code{\link{constellation}()},
\code{\link{get_attribute_names}()},
\code{\link{get_measure_names}()},
\code{\link{get_role_playing_dimension_names}()},
\code{\link{get_similar_attribute_values}()},
\code{\link{replace_attribute_values}()},
\code{\link{role_playing_dimension}()},
\code{\link{set_attribute_names}()},
\code{\link{set_measure_names}()},
\code{\link{snake_case}()},
\code{\link{star_database}()}
}
\concept{star database and constellation definition functions}
