% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_flat_table}
\alias{interpret_operation_flat_table}
\title{Interpret operation}
\usage{
interpret_operation_flat_table(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
Interpret operation
}
\keyword{internal}
