% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_join_lookup_table}
\alias{interpret_operation_join_lookup_table}
\title{Interpret operation}
\usage{
interpret_operation_join_lookup_table(ft, op, lookup_tables, file, last_op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{lookup_tables}{lookup tables}

\item{file}{file to write the code}

\item{last_op}{A boolean, is the last operation?}
}
\value{
A flat table.
}
\description{
operation,          name,          details, details2
"join_lookup_table", fk_attributes, pos)
}
\keyword{internal}
