% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_remove_instances_without_measures}
\alias{interpret_operation_remove_instances_without_measures}
\title{Interpret operation}
\usage{
interpret_operation_remove_instances_without_measures(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,
"remove_instances_without_measures")
}
\keyword{internal}
