% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_replace_string}
\alias{interpret_operation_replace_string}
\title{Interpret operation}
\usage{
interpret_operation_replace_string(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,       name,       details, details2
"replace_string", attributes, string, replacement)
}
\keyword{internal}
