% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_role_playing_dimension}
\alias{interpret_operation_role_playing_dimension}
\title{Interpret operation}
\usage{
interpret_operation_role_playing_dimension(ft, op, file, last_op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{file}{file to write the code}

\item{last_op}{A boolean, is the last operation?}
}
\value{
A flat table.
}
\description{
operation,                name, details, details2
"role_playing_dimension",  rpd,  roles,   att_names)
}
\keyword{internal}
