% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{interpret_operation_select_attributes}
\alias{interpret_operation_select_attributes}
\title{Interpret operation}
\usage{
interpret_operation_select_attributes(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,          name
"select_attributes", attributes)
}
\keyword{internal}
