% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_select_instances}
\alias{interpret_operation_select_instances}
\title{Interpret operation}
\usage{
interpret_operation_select_instances(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,       name,       details, details2
"select_instances", not, attributes, unlist(values)
}
\keyword{internal}
