% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_separate_measures}
\alias{interpret_operation_separate_measures}
\title{Interpret operation}
\usage{
interpret_operation_separate_measures(ft, op, sel_measure_group)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}

\item{sel_measure_group}{measure group index (to return)}
}
\value{
A flat table.
}
\description{
operation,          name,     details, details2
"separate_measures", measures, names, na_rm)
}
\keyword{internal}
