% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_transform_from_values}
\alias{interpret_operation_transform_from_values}
\title{Interpret operation}
\usage{
interpret_operation_transform_from_values(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
operation,              name,
"transform_from_values", attribute)
}
\keyword{internal}
