% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{interpret_operation_transform_to_measure}
\alias{interpret_operation_transform_to_measure}
\title{Interpret operation}
\usage{
interpret_operation_transform_to_measure(ft, op)
}
\arguments{
\item{ft}{flat table}

\item{op}{operation}
}
\value{
A flat table.
}
\description{
Interpret operation
}
\keyword{internal}
