% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances_table.R
\name{prepare_to_join}
\alias{prepare_to_join}
\title{Transform the instances table implemented by a \code{tibble} to join}
\usage{
prepare_to_join(instances, unknown_value)
}
\arguments{
\item{instances}{A \code{tibble}, the instances table.}

\item{unknown_value}{A string, value used to replace NA values in dimensions.}
}
\value{
A \code{tibble}.
}
\description{
Transform all fields in the instances table to character type and replace
the \code{NA} values to facilitate the join operation.
}
\keyword{internal}
