% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{refresh_deployments}
\alias{refresh_deployments}
\title{Refresh deployments}
\usage{
refresh_deployments(db)
}
\arguments{
\item{db}{A \code{star_database} object.}
}
\value{
A \code{star_database} object.
}
\description{
Generate sql code for the first refresh operation.
}
\keyword{internal}
