% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{replace_empty_values_table}
\alias{replace_empty_values_table}
\title{Replace empty values with the unknown value}
\usage{
replace_empty_values_table(ft, attributes = NULL)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{attributes}{A vector of names.}
}
\value{
A \code{flat_table} object.
}
\description{
Replace empty values with the unknown value
}
\keyword{internal}
