% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_table.R
\name{select_measures}
\alias{select_measures}
\alias{select_measures.flat_table}
\title{Select measures of a flat table}
\usage{
select_measures(ft, measures)

\method{select_measures}{flat_table}(ft, measures)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{measures}{A vector of names.}
}
\value{
A \code{flat_table} object.
}
\description{
Select only the indicated measures from the flat table.
}
\examples{

ft <- flat_table('iris', iris) |>
  select_measures(measures = c('Sepal.Length', 'Sepal.Width'))

}
\seealso{
\code{\link{flat_table}}

Other flat table definition functions: 
\code{\link{flat_table}()},
\code{\link{read_flat_table_file}()},
\code{\link{read_flat_table_folder}()},
\code{\link{select_attributes}()}
}
\concept{flat table definition functions}
