% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/role_playing.R
\name{share_dimensions}
\alias{share_dimensions}
\title{Share the given dimensions in the database}
\usage{
share_dimensions(db, dims)
}
\arguments{
\item{db}{\code{star_database} object.}

\item{dims}{Vector of dimension names.}
}
\value{
A \code{star_database} object.
}
\description{
Share the given dimensions in the database
}
\keyword{internal}
