% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{star_database}
\alias{star_database}
\title{\code{star_database} S3 class}
\usage{
star_database(schema, instances)
}
\arguments{
\item{schema}{A \code{star_schema} object.}

\item{instances}{A flat table to define the database instances according to the schema.}
}
\value{
A \code{star_database} object.
}
\description{
An \code{star_database} object is created from a \code{star_schema} object and a flat
table that contains the data from which database instances are derived.
}
\details{
Measures and attributes of the \code{star_schema} must correspond to the names of
the columns of the flat table.
}
\examples{

dput(colnames(ft_num)) # ft_num is defined in the package.
# Measures must be numerical

s <- star_schema() |>
  define_facts(fact_schema(
    name = "mrs_cause",
    measures = c(
      "Pneumonia and Influenza Deaths",
      "All Deaths"
    )
  )) |>
  define_dimension(dimension_schema(
    name = "when",
    attributes = c(
      "Week Ending Date",
      "WEEK",
      "Year"
    )
  )) |>
  define_dimension(dimension_schema(
    name = "where",
    attributes = c(
      "REGION",
      "State",
      "City"
    )
  ))

db <- star_database(s, ft_num)

}
\seealso{
\code{\link{fact_schema}}

Other star schema definition functions: 
\code{\link{define_dimension}()},
\code{\link{define_facts}()},
\code{\link{dimension_schema}()},
\code{\link{fact_schema}()},
\code{\link{star_schema}()}
}
\concept{star schema definition functions}
