% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_database.R
\name{star_database}
\alias{star_database}
\title{\code{star_database} S3 class}
\usage{
star_database(schema, instances)
}
\value{
A \code{star_database} object.
}
\description{
An \code{star_database} object is created from a \code{star_schema} object and a flat
table that contains the data from which database instances are derived.
}
\details{
Measures and attributes of the \code{star_schema} must correspond to the names of
the columns of the flat table.
}
\examples{

s <- star_schema()

}
\seealso{
\code{\link{fact_schema}}

Other star schema definition functions: 
\code{\link{define_dimension}()},
\code{\link{define_facts}()},
\code{\link{dimension_schema}()},
\code{\link{fact_schema}()},
\code{\link{star_schema}()}
}
\concept{star schema definition functions}
