% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_schema.R
\name{star_schema}
\alias{star_schema}
\title{\code{star_schema} S3 class}
\usage{
star_schema()
}
\value{
A \code{star_schema} object.
}
\description{
An empty \code{star_schema} object is created in which definition of facts
and dimensions can be added.
}
\details{
To get a star database (a \code{star_database} object) we need a flat table
and a \code{star_schema} object. The definition of facts and dimensions in
the \code{star_schema} object is made from the flat table columns.
}
\examples{

s <- star_schema()

}
\seealso{
\code{\link{star_database}}

Other star schema definition functions: 
\code{\link{define_dimension}()},
\code{\link{define_facts}()},
\code{\link{dimension_schema}()},
\code{\link{fact_schema}()}
}
\concept{star schema definition functions}
