% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoattribute.R
\name{summarize_layer}
\alias{summarize_layer}
\title{Summarize geometry of a layer}
\usage{
summarize_layer(layer, attribute)
}
\arguments{
\item{layer}{A \code{sf} object.}

\item{attribute}{A string, attribute name.}
}
\value{
A \code{sf} object.
}
\description{
Groups the geometric elements of a layer according to the values of the indicated
attribute.
}
\examples{

file <-
  system.file(
    "extdata",
    "layer_us_level.gpkg",
    package = "rolap"
  )
layer_us_state <- sf::st_read(file, 'layer_us_state',
                              quiet = TRUE)

layer <-
  summarize_layer(layer = layer_us_state, attribute = "REGION")

}
\seealso{
Other star database geographic attributes: 
\code{\link{define_geoattribute}()},
\code{\link{get_geometry}()},
\code{\link{get_point_geometry}()},
\code{\link{get_unrelated_instances}()}
}
\concept{star database geographic attributes}
