% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactions_db}
\alias{transactions_db}
\title{Czech debit card company specialising on payments at gas stations (tables)}
\format{
A \code{star_database}.
}
\source{
\url{https://relational.fit.cvut.cz/dataset/CCS}
}
\usage{
transactions_db
}
\description{
Data base in \code{dm} format with a summary from the data available at the source.
}
\details{
The development has been done in the vignette \code{vignette("v20-rdbms-dm")}.
}
\seealso{
Other database example data: 
\code{\link{db_finest}},
\code{\link{db_summary}}
}
\concept{database example data}
\keyword{datasets}
