% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{update_according_to}
\alias{update_according_to}
\alias{update_according_to.flat_table}
\title{Update a flat table according to another structure}
\usage{
update_according_to(ft, ob_op, star, sel_measure_group)

\method{update_according_to}{flat_table}(ft, ob_op, star = 1, sel_measure_group = 1)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{ob_op}{An object with defined modification operations.}

\item{star}{Star database name or index in constellation.}

\item{sel_measure_group}{A vector of integers, if measures are separated into
groups, indicate which group to consider.}
}
\value{
A \code{flat_table} or \code{star_database} object.
}
\description{
Update a flat table with the operations of another structure based on a flat
table.
}
\examples{

ft <- flat_table('iris', iris)

ft <- flat_table('ft_num', ft_num)

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{remove_instances_without_measures}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{replace_unknown_values}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
