% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{update_according_to}
\alias{update_according_to}
\alias{update_according_to.flat_table}
\title{Update a flat table according to another structure}
\usage{
update_according_to(
  ft,
  ob,
  return_flat_table,
  begin_in_star_database,
  star,
  sel_measure_group,
  out_file,
  function_name,
  only_show_function
)

\method{update_according_to}{flat_table}(
  ft,
  ob,
  return_flat_table = FALSE,
  begin_in_star_database = FALSE,
  star = 1,
  sel_measure_group = 1,
  out_file = NULL,
  function_name = "transform_instance_table",
  only_show_function = FALSE
)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{ob}{A \code{flat_table} or \code{star_database} object with defined modification
operations.}

\item{return_flat_table}{Boolean, even if there are more operations, do not
transform the flat table into a star database.}

\item{begin_in_star_database}{Boolean, start with star database creation
operation.}

\item{star}{A string, star database name or index in constellation.}

\item{sel_measure_group}{A vector of integers, if measures are separated into
groups, indicate which group to consider.}

\item{out_file}{A string, name of the file in which to store the update
instructions that are applied.}

\item{function_name}{A string, name of the function to generate in the file.}

\item{only_show_function}{A boolean, return a vector with the function lines.}
}
\value{
A \code{flat_table} or \code{star_database} object.
}
\description{
Update a flat table with the operations of another structure based on a flat
table.
}
\examples{

f1 <- flat_table('ft_num', ft_cause_rpd) |>
  as_star_database(mrs_cause_schema_rpd) |>
  replace_attribute_values(
    name = "When Available",
    old = c('1962', '11', '1962-03-14'),
    new = c('1962', '3', '1962-01-15')
  ) |>
  group_dimension_instances(name = "When")
f2 <- flat_table('ft_num2', ft_cause_rpd) |>
  update_according_to(f1)

}
\seealso{
\code{\link{flat_table}}

Other flat table transformation functions: 
\code{\link{add_custom_column}()},
\code{\link{remove_instances_without_measures}()},
\code{\link{replace_empty_values}()},
\code{\link{replace_string}()},
\code{\link{replace_unknown_values}()},
\code{\link{select_attributes}()},
\code{\link{select_instances_by_comparison}()},
\code{\link{select_instances}()},
\code{\link{select_measure_group}()},
\code{\link{select_measures}()},
\code{\link{separate_measures}()},
\code{\link{transform_attribute_format}()},
\code{\link{transform_from_values}()},
\code{\link{transform_to_attribute}()},
\code{\link{transform_to_measure}()},
\code{\link{transform_to_values}()}
}
\concept{flat table transformation functions}
