% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_according_to.R
\name{update_according_to}
\alias{update_according_to}
\alias{update_according_to.flat_table}
\title{Update a flat table according to another structure}
\usage{
update_according_to(ft, sdb, star, sdb_operations)

\method{update_according_to}{flat_table}(ft, sdb, star = 1, sdb_operations = NULL)
}
\arguments{
\item{ft}{A \code{flat_table} object.}

\item{sdb}{A \code{star_database} object with defined modification operations.}

\item{star}{A string or integer, star database name or index in constellation.}

\item{sdb_operations}{A \code{star_database} object with new defined modification
operations.}
}
\value{
A \code{star_database_update} object.
}
\description{
Update a flat table with the operations of another structure based on a flat
table.
}
\examples{

f1 <- flat_table('ft_num', ft_cause_rpd) |>
  as_star_database(mrs_cause_schema_rpd) |>
  replace_attribute_values(
    name = "When Available",
    old = c('1962', '11', '1962-03-14'),
    new = c('1962', '3', '1962-01-15')
  ) |>
  group_dimension_instances(name = "When")
f2 <- flat_table('ft_num2', ft_cause_rpd) |>
  update_according_to(f1)

}
\seealso{
Other star database refresh functions: 
\code{\link{get_existing_fact_instances}()},
\code{\link{get_new_dimension_instances}()}
}
\concept{star database refresh functions}
